Move app::Key to its own key.cpp file
